rename cache service to a common name and add it's tests
